Bubbled out a OpenWeb object to open links
