+++ src/main/scala/dbtarzan/gui/OpenWeb.scala
@@ -1,0 +1,19 @@
+package dbtarzan.gui
+
+import dbtarzan.gui.Main.hostServices
+
+object OpenWeb {
+  def openWeb(url: String): Unit = openWebTry1(url)
+
+  private def openWebTry1(url: String): Unit = try {
+    val p = new ProcessBuilder("xdg-open", url).start()
+    if (!p.isAlive)
+      openWebTry2(url)
+  }
+  catch {
+    case e: Throwable => openWebTry2(url)
+  }
+
+  private def openWebTry2(url: String): Unit =
+    hostServices.showDocument(url)
+}
